ref(rust): rename runtime_config module to sentry_options
The module now reads exclusively from the sentry-options store, so name
it accordingly. Because the crate-root module name shadows the external
`sentry_options` crate, `init_sentry_options` in lib.rs disambiguates its
extern-crate references with a leading `::`.
