+++ rust_snuba/src/lib.rs
@@ -7,12 +7,15 @@ pub(crate) const SNUBA_SCHEMA: &str =
 /// (now deprecated) `sentry_options::init_with_schemas`. This keeps repeated
 /// calls -- e.g. across unit tests sharing one process -- from erroring with
 /// `AlreadyInitialized`.
-pub(crate) fn init_sentry_options() -> sentry_options::Result<()> {
-    match sentry_options::Options::builder()
+// `::sentry_options` (leading `::`) refers to the external crate; a bare
+// `sentry_options` here would resolve to the crate-local `sentry_options` module
+// declared below, which shadows the extern crate name at the crate root.
+pub(crate) fn init_sentry_options() -> ::sentry_options::Result<()> {
+    match ::sentry_options::Options::builder()
         .with_schemas(&[("snuba", SNUBA_SCHEMA)])
         .init()
     {
-        Err(sentry_options::OptionsError::AlreadyInitialized) => Ok(()),
+        Err(::sentry_options::OptionsError::AlreadyInitialized) => Ok(()),
         other => other,
     }
 }
@@ -25,7 +28,7 @@ mod logging;
 mod metrics;
 mod processors;
 mod rebalancing;
-mod runtime_config;
+mod sentry_options;
 mod strategies;
 mod types;
